Organize Chrome extension files
