+++ README-ja.md
@@ -24,7 +24,7 @@
 
    * Chrome を開き、アドレスバーに `chrome://extensions/` を入力
    * 右上の **デベロッパーモード** を有効化
-   * **パッケージ化されていない拡張機能を読み込む** をクリックし、このディレクトリを選択
+   * **パッケージ化されていない拡張機能を読み込む** をクリックし、`chrome` ディレクトリを選択
 3. ツールバーのタイマーアイコンをクリックしてポップアップを開き、カウントダウンを開始
 
 ## 使い方
