+++ chrome/formatTime.js
