+++ chrome/manifest.json
@@ -7,9 +7,9 @@
         "storage"
     ],
     "icons": {
-        "16": "icons/icon16.png",
-        "48": "icons/icon48.png",
-        "128": "icons/icon128.png"
+        "16": "../icons/icon16.png",
+        "48": "../icons/icon48.png",
+        "128": "../icons/icon128.png"
     },
     "action": {
         "default_popup": "popup.html"
