+++ chrome/package.json
