+++ chrome/popup.css
