+++ chrome/popup.js
